Modify operator
