+++ src/Web/ClientApp/src/inspection/InspectionItemOperator.tsx
@@ -8,7 +8,7 @@ import InspectionItemReducer, {
 export const InspectionItemOperator = () => {
   const [inspectionItem, dispatch] = useReducer(InspectionItemReducer, {
     inspection_item_id: Math.random().toString(36).substr(2, 9),
-    inspection_content: "",
+    inspection_content: '',
     input_type: 1,
     choices: [],
   });
@@ -37,6 +37,7 @@ export const InspectionItemOperator = () => {
     addChoice, removeChoice, updateChoice,
     setChoices,
   ];
+
 }
 
 /**
