+++ src/Web/ClientApp/src/inspection/Types.tsx
@@ -55,3 +55,14 @@ export type InspectionSheetAction = {
   };
 };
 
+export interface InspectionSheetContextType {
+  inspectionSheet: InspectionSheet;
+  setSheet: (sheet: InspectionSheet) => void;
+  updateField: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  addEquipment: () => void,
+  removeEquipment: (id: string) => void,
+  updateEquipment: (event: React.ChangeEvent<HTMLInputElement>, id: string) => void,
+  addInspectionItem: (id: string, item: InspectionItem) => void,
+  removeInspectionItem: (id: string, itemId: string) => void,
+  updateInspectionItem: (id: string, item: InspectionItem) => void,
+};
